py/data/pillow-notes/pyproject.toml
[project]
name = 'pillow-notes'
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["matplotlib>=3.9.2", "numpy>=2.1.3", "pillow>=11.0.0"]

[tool.uv]
dev-dependencies = ["ruff", "pylint"]

[tool.ruff.lint]
extend-select = [
    "I",
    "W",
    "E",
    "D",
    "UP",
    "B",
    "SIM",
    "F",
    "S",
    "A",
    "C4",
    "ISC",
    "TCH",
    "PTH",
    "PL",
    "PERF",
    "FURB",
    # "RUF",
]

[tool.pyright]
typeCheckingMode = "basic"
